Added basic item list api

+++ ufod/crates/ds-core/src/api/meta.rs
@@ -121,7 +121,7 @@ where
 	async fn get_items(
 		&self,
 		class: ClassHandle,
-		page_size: usize,
-		start_at: usize,
+		page_size: u32,
+		start_at: u32,
 	) -> Result<Vec<ItemData>, MetastoreError>;
 }
